+++ core-ng/src/test/java/core/framework/api/util/StringsTest.java
@@ -2,6 +2,7 @@ package core.framework.api.util;
 
 import org.junit.Test;
 
+import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
@@ -40,5 +41,18 @@ public class StringsTest {
 
         assertFalse(Strings.isEmpty("1"));
         assertFalse(Strings.isEmpty(" 1"));
+    }
+
+    @Test
+    public void split() {
+        assertArrayEquals(new String[]{""}, Strings.split("", '/'));
+        assertArrayEquals(new String[]{"", ""}, Strings.split("/", '/'));
+        assertArrayEquals(new String[]{"", "", ""}, Strings.split("//", '/'));
+        assertArrayEquals(new String[]{"", "1"}, Strings.split("/1", '/'));
+        assertArrayEquals(new String[]{"", "1", ""}, Strings.split("/1/", '/'));
+        assertArrayEquals(new String[]{"1", ""}, Strings.split("1/", '/'));
+        assertArrayEquals(new String[]{"1", "2"}, Strings.split("1/2", '/'));
+        assertArrayEquals(new String[]{"1", "2", "", "3"}, Strings.split("1/2//3", '/'));
+        assertArrayEquals(new String[]{"1", "2", "3"}, Strings.split("1/2/3", '/'));
     }
 }